Refactor routing structure by reintroducing AppRoutes component and updating Sidebar import paths
